comments are fixed
